Move isDevLocal and other variables into /lib/env.
