+++ apps/app/src/lib/auth/index.ts
@@ -1,5 +1,5 @@
 import type { AuthService } from "./types";
-import { isStackAuthEnabled } from "~/lib/utils";
+import { isStackAuthEnabled } from "~/lib/env";
 
 // FUTURE: use dynamic imports to enable tree shaking
 import * as stack from "./stackAuth";
