+++ apps/app/src/lib/auth/stackAuth.ts
@@ -2,6 +2,7 @@ import { useRouter } from "next/navigation";
 
 import type { AuthService } from "./types";
 import { userStore } from "~/stores/userStore";
+import { STACK_PROJECT_ID, STACK_PUBLISHABLE_CLIENT_KEY } from "~/lib/env";
 
 import { StackClientApp } from "@stackframe/react";
 
@@ -10,9 +11,8 @@ let _stackApp: StackClientApp<true, string> | undefined;
 function getStackApp(): StackClientApp<true, string> {
   if (!_stackApp) {
     _stackApp = new StackClientApp({
-      projectId: process.env.NEXT_PUBLIC_STACK_PROJECT_ID!,
-      publishableClientKey:
-        process.env.NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY!,
+      projectId: STACK_PROJECT_ID,
+      publishableClientKey: STACK_PUBLISHABLE_CLIENT_KEY,
       tokenStore: "cookie", // Client-side cookies
       urls: {
         signIn: "/signin",
