+++ apps/app/src/lib/utils.ts
@@ -1,8 +1,3 @@
-export const isDev = process.env.NODE_ENV === "development";
-export const isStackAuthEnabled = !!process.env.NEXT_PUBLIC_STACK_PROJECT_ID;
-
-export const isDevLocal = !process.env.NEXT_PUBLIC_API_URL;
-
 export function getCookie(name: string): string | undefined {
   if (typeof document !== "undefined") {
     const value = document.cookie
